update motion detection

+++ include/basetype.h
@@ -8,14 +8,14 @@
 #ifndef __BASETYPE_H__
 #define __BASETYPE_H__
 
-typedef signed char     s8;
-typedef signed short    s16;
-typedef signed int      s32;
-typedef signed long     s64;
+typedef signed char         s8;
+typedef signed short        s16;
+typedef signed int          s32;
+typedef signed long long    s64;
 
-typedef unsigned char     u8;
-typedef unsigned short    u16;
-typedef unsigned int      u32;
-typedef unsigned long     u64;
+typedef unsigned char       u8;
+typedef unsigned short      u16;
+typedef unsigned int        u32;
+typedef unsigned long long  u64;
 
 #endif
